app/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
mysql-connector-python==8.3.0
pydantic==2.6.4
PyJWT==2.8.0
python-multipart==0.0.9
google-cloud-storage==2.16.0
google-auth==2.29.0
